intをほぼuint32_tに変更

+++ sugiEngine/engine/audio/Sound.h
@@ -28,7 +28,7 @@ struct SoundData {
 	//バッファの先頭アドレス
 	BYTE* pBuffer;
 	//バッファのサイズ
-	unsigned int bufferSize;
+	size_t bufferSize;
 };
 
 class Sound
