+++ sugiEngine/engine/light/CircleShadow.h
@@ -18,7 +18,7 @@ public:
 		XMFLOAT3 atten;
 		float pad3;
 		XMFLOAT2 factorAngleCos;
-		unsigned int active;
+		size_t active;
 		float pad4;
 	};
